+++ app/src/main/java/android/skills_market/ResponsesModel.kt
@@ -1,0 +1,6 @@
+package android.skills_market
+
+data class ResponsesModel(
+    val imageId:Int,
+    val title:String
+)
